% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PerformMeanFilterByChr}
\alias{PerformMeanFilterByChr}
\title{PerformMeanFilterByChr()}
\usage{
PerformMeanFilterByChr(object)
}
\arguments{
\item{object}{casper object}
}
\value{
object
}
\description{
Recusive iterative median filtering is applied for each chromosome
}
